split timeseries anomalies to anomalies and errors
