+++ orion/pipelines/verified/arima/arima.json
@@ -6,7 +6,7 @@
         "mlprimitives.custom.timeseries_preprocessing.rolling_window_sequences",
         "numpy.reshape",
         "statsmodels.tsa.arima_model.Arima",
-        "orion.primitives.timeseries_anomalies.regression_errors",
+        "orion.primitives.timeseries_errors.regression_errors",
         "orion.primitives.timeseries_anomalies.find_anomalies"
     ],
     "init_params": {
