+++ orion/pipelines/verified/dense_autoencoder/dense_autoencoder.json
@@ -5,7 +5,7 @@
         "sklearn.preprocessing.MinMaxScaler",
         "mlprimitives.custom.timeseries_preprocessing.rolling_window_sequences",
         "keras.Sequential.DenseSeq2Seq",
-        "orion.primitives.timeseries_anomalies.reconstruction_errors",
+        "orion.primitives.timeseries_errors.reconstruction_errors",
         "orion.primitives.timeseries_anomalies.find_anomalies"
     ],
     "init_params": {
@@ -45,7 +45,7 @@
         "keras.Sequential.DenseSeq2Seq#1": {
             "y": "X"
         },
-        "orion.primitives.timeseries_anomalies.reconstruction_errors#1": {
+        "orion.primitives.timeseries_errors.reconstruction_errors#1": {
             "y": "X"
         }
     }
