+++ orion/pipelines/verified/lstm_autoencoder/lstm_autoencoder_gpu.json
@@ -5,7 +5,7 @@
         "sklearn.preprocessing.MinMaxScaler",
         "mlprimitives.custom.timeseries_preprocessing.rolling_window_sequences",
         "keras.Sequential.LSTMSeq2Seq",
-        "orion.primitives.timeseries_anomalies.reconstruction_errors",
+        "orion.primitives.timeseries_errors.reconstruction_errors",
         "orion.primitives.timeseries_anomalies.find_anomalies"
     ],
     "init_params": {
@@ -84,7 +84,7 @@
         "keras.Sequential.LSTMSeq2Seq#1": {
             "y": "X"
         },
-        "orion.primitives.timeseries_anomalies.reconstruction_errors#1": {
+        "orion.primitives.timeseries_errors.reconstruction_errors#1": {
             "y": "X"
         }
     }
