+++ orion/pipelines/verified/lstm_dynamic_threshold/lstm_dynamic_threshold_gpu.json
@@ -5,7 +5,7 @@
         "sklearn.preprocessing.MinMaxScaler",
         "mlprimitives.custom.timeseries_preprocessing.rolling_window_sequences",
         "keras.Sequential.LSTMTimeSeriesRegressor",
-        "orion.primitives.timeseries_anomalies.regression_errors",
+        "orion.primitives.timeseries_errors.regression_errors",
         "orion.primitives.timeseries_anomalies.find_anomalies"
     ],
     "init_params": {
