+++ setup.py
@@ -28,7 +28,7 @@ install_requires = [
     'scikit-learn>=0.20.1,<0.21',
     'tabulate>=0.8.3,<0.9',
     'Keras>=2.1.6,<2.4',
-    'similaritymeasures>=0.4.3',
+    'pyts>=0.9,<0.11',
     'azure-cognitiveservices-anomalydetector>=0.3,<0.4',
 
     # fix conflict
